feat: move peer connection related code into peer manager
