+++ waku/v2/protocol/peer_exchange/client.go
@@ -10,8 +10,8 @@ import (
 	"github.com/ethereum/go-ethereum/rlp"
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/libp2p/go-msgio/pbio"
-	v2 "github.com/waku-org/go-waku/waku/v2"
 	"github.com/waku-org/go-waku/waku/v2/metrics"
+	"github.com/waku-org/go-waku/waku/v2/peermanager"
 	"github.com/waku-org/go-waku/waku/v2/peerstore"
 	wenr "github.com/waku-org/go-waku/waku/v2/protocol/enr"
 	"github.com/waku-org/go-waku/waku/v2/protocol/peer_exchange/pb"
@@ -104,7 +104,7 @@ func (wakuPX *WakuPeerExchange) handleResponse(ctx context.Context, response *pb
 		go func() {
 			defer wakuPX.wg.Done()
 			for _, p := range discoveredPeers {
-				peer := v2.PeerData{
+				peer := peermanager.PeerData{
 					Origin:   peerstore.PeerExchange,
 					AddrInfo: p.addrInfo,
 					ENR:      p.enr,
