+++ waku/v2/rendezvous/rendezvous_test.go
@@ -14,21 +14,21 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/waku-org/go-waku/tests"
 	"github.com/waku-org/go-waku/waku/persistence/sqlite"
-	v2 "github.com/waku-org/go-waku/waku/v2"
+	"github.com/waku-org/go-waku/waku/v2/peermanager"
 	"github.com/waku-org/go-waku/waku/v2/utils"
 )
 
 type PeerConn struct {
-	ch chan v2.PeerData
+	ch chan peermanager.PeerData
 }
 
-func (p PeerConn) PeerChannel() chan<- v2.PeerData {
+func (p PeerConn) PeerChannel() chan<- peermanager.PeerData {
 	return p.ch
 }
 
 func NewPeerConn() PeerConn {
 	x := PeerConn{}
-	x.ch = make(chan v2.PeerData, 1000)
+	x.ch = make(chan peermanager.PeerData, 1000)
 	return x
 }
